Now max_depth < 0 means unlimited depth

+++ include/DimRed.hpp
@@ -29,7 +29,7 @@ struct Net : torch::nn::Module {
     // For pretraining
     at::Tensor forward(const at::Tensor & x);
     void copy(const std::shared_ptr<Net> & net);
-    void warmstart(const std::string & chk, const size_t & chk_depth);
+    void warmstart(const std::string & chk, const int64_t & chk_depth);
     void freeze(const size_t & freeze);
 };
